+++ lua/jah/hello_ack.lua
@@ -1,25 +1,24 @@
--- @name hello ack
--- @version 0.1.0
--- @author jah
--- @txt ack test
+-- param ack
+--
+-- ack test script
+--
 
-ControlSpec = require 'controlspec'
-Control = require 'control'
-Scroll = require 'jah/scroll'
-Formatters = require 'jah/formatters'
+local ControlSpec = require 'controlspec'
+local Control = require 'control'
+local Formatters = require 'jah/formatters'
+local FS = require 'fileselect'
 
 engine.name = 'Ack'
 
-local debug = false
-
-local bool_spec = ControlSpec.new(0, 1, 'lin', 1, 0, "")
+local midi_cc_spec = ControlSpec.new(0, 127, 'lin', 1, 0, "")
 
 --[[
 TODO: looping
 local loop_start_spec = ControlSpec.UNIPOLAR
-local loop_end_spec = ControlSpec.new(0, 1, 'lin', 0, 1, "")
+local end_spec = ControlSpec.new(0, 1, 'lin', 0, 1, "")
+local loop_point_spec = ControlSpec.UNIPOLAR
 ]]
-local speed_spec = ControlSpec.new(0, 5, 'lin', 0, 1, "")
+local speed_spec = ControlSpec.new(0.05, 5, 'lin', 0, 1, "")
 -- local slew_spec = ControlSpec.new(0, 5, 'lin', 0, 0, "") -- TODO: enable slews
 local volume_spec = ControlSpec.DB:copy()
 volume_spec.default = -10
@@ -40,71 +39,6 @@ local delay_time_spec = ControlSpec.new(0.0001, 5, 'exp', 0, 0.1, "secs")
 local reverb_room_spec = ControlSpec.new(0, 1, 'lin', 0, 0.5, "")
 local reverb_damp_spec = ControlSpec.new(0, 1, 'lin', 0, 0.5, "")
 
-local midi_in = Control.new("midi in", bool_spec, Formatters.unipolar_as_enabled_disabled)
-midi_in:set(1) -- TODO: hack, better idea: encoder deltas configured per param or set_min_mapped_value()/set_max_mapped_value()
-
-local midi_selects = Control.new("midi selects", bool_spec, Formatters.unipolar_as_true_false)
-midi_selects:set(0.98) -- TODO: hack, better idea: encoder deltas configured per param or set_min_mapped_value()/set_max_mapped_value()
-
-local trig_on_change = Control.new("trig on value change", bool_spec, Formatters.unipolar_as_true_false)
-trig_on_change:set_raw(0.98) -- TODO: hack, better idea: encoder deltas configured per param or set_min_mapped_value()/set_max_mapped_value()
-
-local channel_params = {}
-for i=0,7 do
-  local p = {}
---[[
-TODO: looping
-  p.loop_start = Control.new((i+1)..": loop start", loop_start_spec, Formatters.unipolar_as_percentage)
-  p.loop_start.action = function(value) engine.loopStart(i, value) end
-  p.loop_end = Control.new((i+1)..": loop end", loop_end_spec, Formatters.unipolar_as_percentage)
-  p.loop_end.action = function(value) engine.loopEnd(i, value) end
-]]
-  p.speed = Control.new((i+1)..": speed", speed_spec, Formatters.unipolar_as_percentage)
-  p.speed.action = function(value) engine.speed(i, value) end
-  p.volume = Control.new((i+1)..": vol", volume_spec, Formatters.std)
-  p.volume.action = function(value) engine.volume(i, value) end
-  p.volume_env_attack = Control.new((i+1)..": vol env atk", volume_env_attack_spec, Formatters.secs_as_ms)
-  p.volume_env_attack.action = function(value) engine.volumeEnvAttack(i, value) end
-  p.volume_env_release = Control.new((i+1)..": vol env rel", volume_env_release_spec, Formatters.secs_as_ms)
-  p.volume_env_release.action = function(value) engine.volumeEnvRelease(i, value) end
-  p.pan = Control.new((i+1)..": pan", ControlSpec.pan(), Formatters.bipolar_as_pan_widget)
-  p.pan.action = function(value) engine.pan(i, value) end
-  p.filter_cutoff = Control.new((i+1)..": filter cutoff", filter_cutoff_spec, Formatters.round(0.001))
-  p.filter_cutoff.action = function(value) engine.filterCutoff(i, value) end
-  p.filter_res = Control.new((i+1)..": filter res", filter_res_spec, Formatters.unipolar_as_percentage)
-  p.filter_res.action = function(value) engine.filterRes(i, value) end
-  p.filter_mode = Control.new((i+1)..": filter mode", filter_mode_spec, Formatters.std)
-  p.filter_mode.action = function(value) engine.filterMode(i, value) end
-  p.filter_env_attack = Control.new((i+1)..": filter env atk", filter_env_attack_spec, Formatters.secs_as_ms)
-  p.filter_env_attack.action = function(value) engine.filterEnvAttack(i, value) end
-  p.filter_env_release = Control.new((i+1)..": filter env rel", filter_env_release_spec, Formatters.secs_as_ms)
-  p.filter_env_release.action = function(value) engine.filterEnvRelease(i, value) end
-  p.filter_env_mod = Control.new((i+1)..": filter env mod", filter_env_mod_spec, Formatters.unipolar_as_percentage)
-  p.filter_env_mod.action = function(value) engine.filterEnvMod(i, value) end
-  p.delay_send = Control.new((i+1)..": delay send", send_spec, Formatters.std)
-  p.delay_send.action = function(value) engine.delaySend(i, value) end
-  p.reverb_send = Control.new((i+1)..": reverb send", send_spec, Formatters.std)
-  p.reverb_send.action = function(value) engine.reverbSend(i, value) end
-  --[[
-  TODO: enable slews
-  p.speed_slew = Control.new((i+1)..": speed slew", slew_spec, Formatters.std)
-  p.speed_slew.action = function(value) engine.speedSlew(i, value) end
-  p.volume_slew = Control.new((i+1)..": vol slew", slew_spec, Formatters.std)
-  p.volume_slew.action = function(value) engine.volumeSlew(i, value) end
-  p.pan_slew = Control.new((i+1)..": pan slew", slew_spec, Formatters.std)
-  p.pan_slew.action = function(value) engine.panSlew(i, value) end
-  p.filter_cutoff_slew = Control.new((i+1)..": filter cutoff slew", slew_spec, Formatters.std)
-  p.filter_cutoff_slew.action = function(value) engine.filterCutoffSlew(i, value) end
-  p.filter_res_slew = Control.new((i+1)..": filter res slew", slew_spec, Formatters.std)
-  p.filter_res_slew.action = function(value) engine.filterResSlew(i, value) end
-  ]]
-  channel_params[i] = p
-end
-
-local delay_time = Control.new("delay time", delay_time_spec, Formatters.secs_as_ms)
-delay_time.action = function(value) engine.delayTimeL(value) end
--- local delay_time_r = Control.new("delay time r", delay_time_spec, Formatters.std) TODO
--- delay_time_r.action = function(value) engine.delayTimeR(value) end TODO
 local delay_feedback = Control.new("delay feedback", delay_time_spec, Formatters.secs_as_ms)
 delay_feedback:set(0.75)
 delay_feedback.action = function(value) engine.delayFeedback(value) end
@@ -113,235 +47,402 @@ reverb_room.action = function(value) engine.reverbRoom(value) end
 local reverb_damp = Control.new("reverb damp", reverb_damp_spec, Formatters.unipolar_as_percentage)
 reverb_damp.action = function(value) engine.reverbDamp(value) end
 
-local function debug_print(str)
-  if debug then print(str) end
+local selected_channel = 0
+local all_selected = false
+local midinote_indicator_level
+local midicc_indicator_level
+local note_downs = {}
+
+local function screen_update_channels()
+  screen.move(0,16)
+  screen.font_size(8)
+  for channel=0,7 do
+    if note_downs[channel] then
+      screen.level(15)
+    elseif selected_channel == channel or all_selected then
+      screen.level(6)
+    else
+      screen.level(2)
+    end
+    screen.text(channel+1)
+  end
+  if all_selected then
+    screen.level(6)
+  else
+    screen.level(0)
+  end
+  screen.text(" all")
+  screen.update()
 end
 
-local function get_channel_from_string(string) -- TODO
-  -- TODO: ugly hack code- change so that this is not a for loop, as we're only interested in first match
-  words = {}
-  for word in string:gmatch("([0-9]+):") do -- TODO: why doesn't "^([0-9]+):" work here?
-    table.insert(words, word)
+local function screen_update_midi_indicators()
+  --screen.move(125, 20)
+  screen.move(0,60)
+  screen.font_size(8)
+  if midi_available then
+    screen.level(15)
+    screen.text("midi:")
+    screen.text(" ")
+    screen.level(midinote_indicator_level)
+    screen.text("note ")
+    screen.level(midicc_indicator_level)
+    screen.text("cc")
+  else
+    screen.level(3)
+    screen.text("no midi")
   end
-  return tonumber(words[1])
+end
+
+local function channel_from_midinote(midinote)
+  local channel = nil
+  if midinote == 60 then
+    channel = 0
+  elseif midinote == 62 then
+    channel = 1
+  elseif midinote == 64 then
+    channel = 2
+  elseif midinote == 65 then
+    channel = 3
+  elseif midinote == 67 then
+    channel = 4
+  elseif midinote == 69 then
+    channel = 5
+  elseif midinote == 71 then
+    channel = 6
+  elseif midinote == 72 then
+    channel = 7
+  end
+  return channel
 end
 
 local function note_on(note, velocity)
-  debug_print("midi / note_on: "..note..", velocity: "..velocity) -- TODO: replace this with on-screen notification
-  local channel = nil
-  if note == 60 then
-    channel = 0
-  elseif note == 62 then
-    channel = 1
-  elseif note == 64 then
-    channel = 2
-  elseif note == 65 then
-    channel = 3
-  elseif note == 67 then
-    channel = 4
-  elseif note == 69 then
-    channel = 5
-  elseif note == 71 then
-    channel = 6
-  elseif note == 72 then
-    channel = 7
-  end
+  local channel = channel_from_midinote(note)
   if channel then
-    debug_print("midi / channel: "..channel)
-    engine.trig(channel)
-    if midi_selects:get() == 1 then
-      local selected_param = scroll.selected_param
-
-      if selected_param == nil or get_channel_from_string(selected_param.name) ~= channel+1 then -- TODO: make this the proper way
-        scroll:navigate_to_lineno(scroll:lookup_lineno(channel_params[channel].speed))
-        redraw()
+    if not note_downs[channel] then
+      note_downs[channel] = true
+      engine.trig(channel)
+      if params:get("midi selects channel") == 2 then
+        selected_channel = channel
       end
+      screen_update_channels()
     end
   end
 end
 
 local function note_off(note)
-  debug_print("midi / note_off: "..note)
+  local channel = channel_from_midinote(note)
+  if channel then
+    note_downs[channel] = false
+    screen_update_channels()
+  end
 end
 
-local function cc(control, value)
-  debug_print("midi / control: "..control..", value: "..value)
+local function cc_set_control(name, controlspec, value)
+  params:set(name, controlspec:map(midi_cc_spec:unmap(value)))
+end
+
+local function cc_delta_control(name, controlspec, value)
+  local delta
+  if value > 0 and value < 64 then
+    delta = value
+  else
+    delta = value - 128
+  end
+  local value = params:get(name)
+  local value_unmapped = controlspec:unmap(value)
+  local new_unmapped_value = value_unmapped + delta/100
+  params:set(name, controlspec:map(new_unmapped_value))
+end
+
+local function cc(ctl, value)
+  local param
+  local spec
+  if ctl == params:get("filter cutoff cc") then
+    param = "filter cutoff"
+    spec = filter_cutoff_spec
+    abs = params:get("filter cutoff cc type") == 1
+  elseif ctl == params:get("filter res cc") then
+    param = "filter res"
+    spec = filter_res_spec
+    abs = params:get("filter res cc type") == 1
+  elseif ctl == params:get("delay send cc") then
+    param = "delay send"
+    spec = send_spec
+    abs = params:get("delay send cc type") == 1
+  elseif ctl == params:get("reverb send cc") then
+    param = "reverb send"
+    spec = send_spec
+    abs = params:get("reverb send cc type") == 1
+  end
+  if param then
+    if abs then
+      if all_selected then
+        for i=0,7 do
+          cc_set_control((i+1)..": "..param, spec, value)
+        end
+      else
+        cc_set_control((selected_channel+1)..": "..param, spec, value)
+      end
+    else
+      if all_selected then
+        for i=0,7 do
+          cc_delta_control((i+1)..": "..param, spec, value)
+        end
+      else
+        cc_delta_control((selected_channel+1)..": "..param, spec, value)
+      end
+    end
+  end
+end
+
+local function add_ack_params()
+  --[[
+  TODO: looping
+  local loop_start_spec = ControlSpec.UNIPOLAR
+  local end_spec = ControlSpec.new(0, 1, 'lin', 0, 1, "")
+  local loop_point_spec = ControlSpec.UNIPOLAR
+  ]]
+  local speed_spec = ControlSpec.new(0.05, 5, 'lin', 0, 1, "")
+  -- local slew_spec = ControlSpec.new(0, 5, 'lin', 0, 0, "") -- TODO: enable slews
+  local volume_spec = ControlSpec.DB:copy()
+  volume_spec.default = -10
+  local send_spec = ControlSpec.DB:copy()
+  send_spec.default = -60
+  local volume_env_attack_spec = ControlSpec.new(0, 1, 'lin', 0, 0.001, "secs")
+  local volume_env_release_spec = ControlSpec.new(0, 3, 'lin', 0, 3, "secs")
+  local filter_env_attack_spec = ControlSpec.new(0, 1, 'lin', 0, 0.001, "secs")
+  local filter_env_release_spec = ControlSpec.new(0, 3, 'lin', 0, 0.25, "secs")
+  local filter_cutoff_spec = ControlSpec.FREQ:copy()
+  filter_cutoff_spec.default = 20000
+  local filter_res_spec = ControlSpec.UNIPOLAR
+  local filter_mode_spec = ControlSpec.new(0, 1, 'lin', 1, 0)
+  local filter_env_mod_spec = ControlSpec.UNIPOLAR
+
+  local delay_time_spec = ControlSpec.new(0.0001, 5, 'exp', 0, 0.1, "secs")
+  local delay_feedback_spec = ControlSpec.new(0, 1.25, 'lin', 0, 0.5, "")
+  local reverb_room_spec = ControlSpec.new(0, 1, 'lin', 0, 0.5, "")
+  local reverb_damp_spec = ControlSpec.new(0, 1, 'lin', 0, 0.5, "")
+
+  for i=1,8 do
+    params:add_file(i..": sample")
+    params:set_action(i..": sample", function(value)
+      if value ~= "-" then
+        engine.loadSample(i-1, value)
+      end
+    end)
+  --[[
+  TODO: looping
+    params:add_control(i..": start", start_spec, Formatters.unipolar_as_percentage)
+    params:set_action(i..": start", function(value) engine.start(i-1, value) end)
+    params:add_control(i..": end", end_spec, Formatters.unipolar_as_percentage)
+    params:set_action(i..": end", function(value) engine.end(i-1, value) end)
+    params:add_control(i..": loop point", loop_point_spec, Formatters.unipolar_as_percentage)
+    params:set_action(i..": loop point", function(value) engine.loopPoint(i-1, value) end)
+    params:add_option(i..": loop", {"off", "on"})
+    params:set_action(i..": loop", function(value) engine.loop(i-1, value) end)
+  ]]
+    params:add_control(i..": speed", speed_spec, Formatters.unipolar_as_percentage)
+    params:set_action(i..": speed", function(value) engine.speed(i-1, value) end)
+    params:add_control(i..": vol", volume_spec, Formatters.std)
+    params:set_action(i..": vol", function(value) engine.volume(i-1, value) end)
+    params:add_control(i..": vol env atk", volume_env_attack_spec, Formatters.secs_as_ms)
+    params:set_action(i..": vol env atk", function(value) engine.volumeEnvAttack(i-1, value) end)
+    params:add_control(i..": vol env rel", volume_env_release_spec, Formatters.secs_as_ms)
+    params:set_action(i..": vol env rel", function(value) engine.volumeEnvRelease(i-1, value) end)
+    params:add_control(i..": pan", ControlSpec.PAN, Formatters.bipolar_as_pan_widget)
+    params:set_action(i..": pan", function(value) engine.pan(i-1, value) end)
+    params:add_control(i..": filter cutoff", filter_cutoff_spec, Formatters.round(0.001))
+    params:set_action(i..": filter cutoff", function(value) engine.filterCutoff(i-1, value) end)
+    params:add_control(i..": filter res", filter_res_spec, Formatters.unipolar_as_percentage)
+    params:set_action(i..": filter res", function(value) engine.filterRes(i-1, value) end)
+    --[[
+    params:add_control(i..": filter mode", filter_mode_spec, Formatters.std)
+    params:set_action(function(value) engine.filterMode(i-1, value) end)
+    ]]
+    params:add_control(i..": filter env atk", filter_env_attack_spec, Formatters.secs_as_ms)
+    params:set_action(i..": filter env atk", function(value) engine.filterEnvAttack(i-1, value) end)
+    params:add_control(i..": filter env rel", filter_env_release_spec, Formatters.secs_as_ms)
+    params:set_action(i..": filter env rel", function(value) engine.filterEnvRelease(i-1, value) end)
+    params:add_control(i..": filter env mod", filter_env_mod_spec, Formatters.unipolar_as_percentage)
+    params:set_action(i..": filter env mod", function(value) engine.filterEnvMod(i-1, value) end)
+    params:add_control(i..": delay send", send_spec, Formatters.std)
+    params:set_action(i..": delay send", function(value) engine.delaySend(i-1, value) end)
+    params:add_control(i..": reverb send", send_spec, Formatters.std)
+    params:set_action(i..": reverb send", function(value) engine.reverbSend(i-1, value) end)
+    --[[
+    TODO: enable slews
+    params:add_control(i..": speed slew", slew_spec, Formatters.std)
+    params:set_action(i..": speed slew", function(value) engine.speedSlew(i-1, value) end)
+    params:add_control(i..": vol slew", slew_spec, Formatters.std)
+    params:set_action(i..": vol slew", function(value) engine.volumeSlew(i-1, value) end)
+    params:add_control(i..": pan slew", slew_spec, Formatters.std)
+    params:set_action(i..": pan slew", function(value) engine.panSlew(i-1, value) end)
+    params:add_control(i..": filter cutoff slew", slew_spec, Formatters.std)
+    params:set_action(i..": filter cutoff slew", function(value) engine.filterCutoffSlew(i-1, value) end)
+    params:add_control(i..": filter res slew", slew_spec, Formatters.std)
+    params:set_action(i..": filter res slew", function(value) engine.filterResSlew(i-1, value) end)
+    ]]
+  end
+
+  params:add_control("delay time", delay_time_spec, Formatters.secs_as_ms)
+  params:set_action("delay time", engine.delayTime)
+  params:add_control("delay feedback", delay_feedback_spec, Formatters.unipolar_as_percentage)
+  params:set_action("delay feedback", engine.delayFeedback)
+  params:add_control("reverb room", reverb_room_spec, Formatters.unipolar_as_percentage)
+  params:set_action("reverb room", engine.reverbRoom)
+  params:add_control("reverb damp", reverb_damp_spec, Formatters.unipolar_as_percentage)
+  params:set_action("reverb damp", engine.reverbRoom)
 end
 
 init = function()
   screen.aa(1)
   screen.line_width(1.0)
 
-  for channelnum=0,7 do
-    for key, param in pairs(channel_params[channelnum]) do param:bang() end
+  local bool = {"false", "true"}
+  params:add_option("midi in", {"disabled", "enabled"}, 2)
+  params:add_option("midi selects channel", bool, 2)
+
+  local cc_list = {}
+  for i=0,127 do
+    cc_list[i] = i
   end
-  delay_time:bang()
-  delay_feedback:bang()
-  reverb_room:bang()
-  reverb_damp:bang()
+  cc_type = {"abs", "rel"}
+  params:add_option("filter cutoff cc", cc_list, 1)
+  params:add_option("filter cutoff cc type", cc_type)
+  params:add_option("filter res cc", cc_list, 2)
+  params:add_option("filter res cc type", cc_type)
+  params:add_option("delay send cc", cc_list, 3)
+  params:add_option("delay send cc type", cc_type)
+  params:add_option("reverb send cc", cc_list, 4)
+  params:add_option("reverb send cc type", cc_type)
+  -- TODO params:add_option("trig on param change", bool)
 
-  local sampleroot = "/home/pi/dust/audio/hello_ack/"
-  engine.loadSample(0, sampleroot.."XR-20_003.wav")
-  engine.loadSample(1, sampleroot.."XR-20_114.wav")
-  engine.loadSample(2, sampleroot.."XR-20_285.wav")
-  engine.loadSample(3, sampleroot.."XR-20_328.wav")
-  engine.loadSample(4, sampleroot.."XR-20_121.wav")
-  engine.loadSample(5, sampleroot.."XR-20_667.wav")
-  engine.loadSample(6, sampleroot.."XR-20_128.wav")
-  engine.loadSample(7, sampleroot.."XR-20_718.wav")
+  add_ack_params()
 
-  scroll = Scroll.new()
-  scroll:push("ack test script")
-  scroll:push("")
-  scroll:push("# ui")
-  scroll:push("")
-  scroll:push("enc2: scroll ui")
-  scroll:push("enc3: change param value")
-  scroll:push("key2: preview sound")
-  scroll:push("key2+enc3: fine adjust")
-  scroll:push("key3: set value to default")
-  scroll:push("")
-  scroll:push("# global settings")
-  scroll:push("")
-  scroll:push(midi_in)
-  scroll:push("(when enabled midi notes ")
-  scroll:push("c4...c5 from external midi")
-  scroll:push("device will trigger sounds)")
-  scroll:push("")
-  scroll:push(midi_selects)
-  scroll:push("(when true and midi is")
-  scroll:push("enabled incoming midi notes")
-  scroll:push("will select channels)")
-  scroll:push("")
-  scroll:push(trig_on_change)
-  scroll:push("(when true sounds will be")
-  scroll:push("triggered upon change of")
-  scroll:push("channel params)")
-  scroll:push("")
-  scroll:push("# ack")
-  scroll:push("")
-  for i=0,7 do
-    -- TODO scroll:push(channel_params[i].loop_start)
-    -- TODO scroll:push(channel_params[i].loop_end)
-    scroll:push(channel_params[i].speed)
-    scroll:push(channel_params[i].volume)
-    scroll:push(channel_params[i].volume_env_attack)
-    scroll:push(channel_params[i].volume_env_release)
-    scroll:push(channel_params[i].pan)
-    -- TODO scroll:push(channel_params[i].filter_mode)
-    scroll:push(channel_params[i].filter_cutoff)
-    scroll:push(channel_params[i].filter_res)
-    scroll:push(channel_params[i].filter_env_attack)
-    scroll:push(channel_params[i].filter_env_release)
-    scroll:push(channel_params[i].filter_env_mod)
-    scroll:push(channel_params[i].delay_send)
-    scroll:push(channel_params[i].reverb_send)
-    --[[
-    TODO: why aren't slews working?
-    scroll:push(channel_params[i].speed_slew)
-    scroll:push(channel_params[i].volume_slew)
-    scroll:push(channel_params[i].pan_slew)
-    -- TODO scroll:push(channel_params[i].filter_cutoff_slew)
-    scroll:push(channel_params[i].filter_res_slew)
-    ]]
-    scroll:push("")
-
-  end
-  scroll:push(delay_time)
-  -- scroll:push(delay_time_r) TODO
-  scroll:push(delay_feedback)
-  scroll:push("")
-  scroll:push(reverb_room)
-  scroll:push(reverb_damp)
-  scroll:push("")
-  scroll:push("...fin")
+  -- params:read("param_ack.pset")
+  params:bang()
 end
 
 redraw = function()
-  if scroll then
-    scroll:redraw(screen)
-  end
+  screen.clear()
+  screen.aa(1)
+  screen.move(0, 8)
+  screen.font_size(8)
+  screen.level(15)
+  screen.text("ack")
+  screen_update_channels()
+  screen_update_midi_indicators()
   screen.update()
-end
-
-local function trig_if_channel_param(param)
-  if param then
-    local channel = get_channel_from_string(param.name)
-    if channel then
-      engine.trig(channel-1)
-    end
-  end
 end
 
 enc = function(n, delta)
   if n == 1 then
     norns.audio.adjust_output_level(delta)
     return
-  end
-
-  if n == 2 then
-    scroll:navigate(delta)
-    redraw()
-  elseif n == 3 then
-    if scroll.selected_param then
-      --[[
-      local d
-      if key2_down then
-        d = delta/500
-      else
-        d = delta/50
+  elseif n == 2 then
+    local new_selection
+    if delta < 0 then
+      if selected_channel ~= 0 then
+        new_selection = selected_channel - 1
       end
-      ]]
-      local param = scroll.selected_param
-      local prev = param:get()
-      param:delta(delta)
-      if param:get() ~= prev and trig_on_change:get() == 1 then
-        trig_if_channel_param(param)
+    else
+      if selected_channel ~= 7 then
+        new_selection = selected_channel + 1
       end
+    end
+    if new_selection then
+      if note_downs[selected_channel] then
+        note_downs[selected_channel] = false
+      end
+      selected_channel = new_selection
       redraw()
+    end
+  else
+    if all_selected then
+      for i=0,7 do
+        params:delta((i+1)..": speed", delta)
+      end
+    else
+      params:delta((selected_channel+1)..": speed", delta)
     end
   end
 end
 
-key = function(n, z)
-  if z == 1 then
-    if n == 2 then
-      key2_down = true
-      trig_if_channel_param(scroll.selected_param)
-    elseif n == 3 then
-      if scroll.selected_param then
-        local param = scroll.selected_param
-        local prev = param:get()
-        param:set_default()
-        if param:get() ~= prev and trig_on_change:get() == 1 then
-          trig_if_channel_param(param)
-        end
-        redraw()
-      end
-    end
+local function trig_channel(channel)
+  engine.trig(channel)
+  if not note_downs[channel] then
+    note_downs[channel] = true
   end
-  if z == 0 then
-    if n == 2 then
-      key2_down = false
-    elseif n == 3 then
+end
+
+local function reset_channel(channel)
+  if note_downs[channel] then
+    note_downs[channel] = false
+  end
+end
+
+local function newfile(what)
+  if what ~= "-" then
+    engine.loadSample(selected_channel, what)
+  end
+end
+
+key = function(n, z)
+  if n == 2 then
+    if z == 1 then
+      if all_selected then
+        for i=0,7 do trig_channel(i) end
+      else
+        trig_channel(selected_channel)
+      end
+      screen_update_channels()
+    else
+      if all_selected then
+        for i=0,7 do reset_channel(i) end
+      else
+        reset_channel(selected_channel)
+      end
+      screen_update_channels()
     end
+  elseif n == 3 then
+    all_selected = z == 1
+    redraw()
+  elseif n==1 and z==1 then
+    FS.enter("/home/pi/dust", newfile)
   end
 end
 
 cleanup = function()
   norns.midi.event = nil
+  -- params:write("param_ack.pset")
 end
 
-norns.midi.event = function(id, status, data1, data2)
-  if midi_in:get() == 1 then
-    -- TODO print(id, status, data1, data2)
+norns.midi.add = function(id, name, dev)
+  midi_available = true
+  midinote_indicator_level = 3
+  midicc_indicator_level = 3
+  redraw()
+end
+
+norns.midi.remove = function(id)
+  midi_available = false
+  redraw()
+end
+
+norns.midi.event = function(id, data)
+  status = data[1]
+  data1 = data[2]
+  data2 = data[3]
+  if params:get("midi in") == 2 then
     if status == 144 then
+      midinote_indicator_level = math.random(15)
       --[[
       if data1 == 0 then
         return -- TODO: filter OP-1 bpm link oddity, is this an op-1 or norns issue?
       end
       ]]
       note_on(data1, data2)
+      redraw()
     elseif status == 128 then
       --[[
       if data1 == 0 then
@@ -350,7 +451,9 @@ norns.midi.event = function(id, status, data1, data2)
       ]]
       note_off(data1)
     elseif status == 176 then
+      midicc_indicator_level = math.random(15)
       cc(data1, data2)
+      redraw()
     elseif status == 224 then
       bend(data1, data2)
     end
